feat: routes refactor (#115)
We need to be able to delete routes if they're not set. The current 
system by which routes are generated means their names are 
unpredictable

By pre-seeding the map with the expected route keys and null values 
the deletion logic can determine which routes are not present and 
should be deleted. This comes with the caveat that custom listeners 
must be manually deleted by a user or deleted by owner reference on 
cluster deletion.

Contributes to: mhub/qp-planning#4629

Signed-off-by: Samuel Hawker <samuel.hawker@ibm.com>