feat: Structural schema (#120)
One of the requirements of structural schemas is that all objects
have a type. This commit makes our CRDs compliant with this by
adding a type that was missing at the top level.

This is something that the upstream Strimzi project are explicitly
and intentionally omitting because they need to support Kubernetes
1.11 (OpenShift 3.11) which does not support this.

We'll keep this change in our fork only until they drop support
for OpenShift 3.11, at which point the change can be made in both
places. This is something that is already in plan for Strimzi.

Contributes to: mhub/qp-planning#4805

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>