feat: New example Event Streams CR templates (#142)
This commit replaces the previous examples with a new core set.

One is a simple minimal CR called "quickstart", which is the name
that the rest of CP4I is using for their demo/PoC CR templates.
It is a single Kafka broker, single ZK node cluster, with no
auth enabled and no geo-replicator. The Kafka cluster config is
modified to reduce the footprint requirements.

The remaining examples are production cluster specifications, for
a three, five and nine broker cluster. The configuration is based
on the setup described in the performance report for the previous
release of Event Streams.

I've tried to reduce the amount of detail needed in the CR, such
as removing some of the attributes we've had in examples that
already matched the default values.
I've modified the spec to remove the requirement to specify the
number of replicas for geo-rep pods, to make the minimum CR yaml
smaller.

I've also fixed the order of attributes in the Event Streams spec
which had gotten out of date with the actual attributes.

I've left "architecture" at the top as it'll be the next thing
that we remove.

Contributes to: mhub/qp-planning#4622

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>