fix: Check if Kafka has a Ready condition (#178)
The previous implementation had a hard-coded assumption that the
ready condition would always be the first condition in the list.
This isn't a safe assumption, as other warnings and conditions
can be included in the list.

This commit makes our check a bit more flexible, so that we can
cope with no list of conditions, an empty list, a list with other
warnings in, etc.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>