fix: Only query replicator dest side users if replicator enabled (#176)
The setClientAuthForReplicator queries the destination side
kafka users if client auth is enabled.  However it does this regardless
of whether replication is enabled.  And the users are only made if
replication is enabled.

So, the case where client auth is on, but replication is not enabled
it needs to not query the destination side kafka users.

Contributes to: mhub/qp-planning#5063

Signed-off-by: Emma Humber <emma.humber@uk.ibm.com>
Signed-off-by: Gavin Royce Quadros <gavin.r.quadros@ibm.com>