feat: Convert Admin REST To Endpoints (#177)
Convert Admin REST to use new Endpoints Model. This will allow users
to configure the endpoints that they want to be created with a
specific configuration. Users can specify what type of service they
want to create, the authentication mechanisms, if the endpoint is TLS,
and what port to access. It will also reconcile the certificates for
each endpoint.

Note that NodePort Certificate Generation is currently not
implemented. Only the route name is put into the SANS if using an
external route otherwise it will put in the service.

Contributes to: mhub/qp-planning#4990

Signed-off-by: Julian Goh <julian.goh@ibm.com>