fix: Enable the rest producer to work with security
Update the image and envars / mounts etc. for the REST producer
to work properly.

Add kafka extension network policy. Need to extend the Kafka network
policy to allow certain traffic to the runas port.

This was in Strimzi code. Have moved into Event Streams code and
added a rule for the REST producer.

Set the owner reference.

Test the new netpol is created.

Contributes to: mhub/qp-planning#4763

Signed-off-by: Chris Patmore <christopher.patmore@ibm.com>