feat: Add ability to set TLS version & rename to internalTls (#207)
Add the ability for the operator to configure Admin REST, REST
Producer, Schema Registry, Collector and UI with a TLS Version
variable. This will be used to drive config to configure the
specified TLS Version for an endpoint or per component.

Also fixed an issue where the operator tried to create the same
service multiple times. Also renamed encryption to internalTls

Contributes to: mhub/qp-planning#5115

Signed-off-by: Julian Goh <julian.goh@ibm.com>