feat: Implement validation of plain listeners (#199)
Check the configuration of EventStreams security and if it is NONE
then make sure that the Kafka Plain Listener is configured WITHOUT
security. This will prevent the situation where REST can't talk to
Kafka due to a lack of a Plain Listener/it can't authenticate with
Kafka.

Contributes to: mhub/qp-planning#4899

Signed-off-by: Julian Goh <julian.goh@ibm.com>