feat: New operator for geo-replicator clusters (#238)
When a new georeplicator is added, if admin-api updates the MM2 CR to
add in the new config then the ES operator will splat over the changes
with what was set originally in the Operator config and all the new
config is lost.

Strimzi mm2 operator attempts to honour the changes made by REST and
recreates the MM2 pod with the new settings, before recreating it
again without them after the ES operator changes them.

Really, REST needs to update the ES CR, not the mm2 cr as that is
where we define the config, but this is pretty risking as patching our
main, top level CR each time a mm2 is made isn't wise (what if we
patch it badly....)

So, as discussed with Sam (and a tiny bit with Dale), separating out
the Replicator part of the ES operator into its own resource that we
can then patch from admin-api.

REST can then make changes to the georeplicator cr, the operator then
updates the MM2 cr, and strimzi deals with the update and we should,
in theory, be all good.

A few items to help with the review

In abstract model I needed to change the following so that the
setAbstractMethod works for both ES and ESGeorep instances

protected void setOwnerReference(EventStreams instance) {
protected void setOwnerReference(CustomResource instance) {

I've made a separate ReplicatorSecretModel, as that's the only bit of
the replicator that's needed by the ES operator (admin api always
needs the secret created).

I've split ReplicatorUserModel into source and destination users.
Source is used by the ES Operator (ie always needed if security is on)
and Dest is used by the ESGR operator.

I've pulled out an AbstractStatus as both ES and ESG crs will need
access to this.

I've removed the georep security on/off envs in Admin-api, I don't
think they are needed.

Status and ESGeorep operator unit tests are coming later, although a
good chunk of the code is still unit tested by the existing
Replicator model tests - I'll add the georep with no admin api test
back in there.

I need to add in a check to ensure the es georep cr name and the name
in the label used to indicate the owning instance match.

I need to work out what to do to replace the GEOREP_ENABLED environment
variable in admin-api, as the ES operator doesn't know the answer to
that - possibly add a value in the georep secret and mount it as an
env? Will do that in a follow up PR too.

Signed-off-by: Emma Humber <emma.humber@uk.ibm.com>