fix: Change owner of operator supporting resources to be clusterwide (#252)
Cluster-scoped Kubernetes resources should not have owner-references
set to namespace-scoped Kubernetes resources. We have heard that this
can cause problems with Kubernetes garbage collection.

This commit changes the owner references for resources created by the
operator init container to be the cluster role created for the
operator deployment.

The cluster role is named dynamically, so the script needs to discover
it. It uses the assumption that there will only one operator running
in a namespace, so it looks for the only cluster role that is bound
to the service account in that namespace.

In the (unlikely?) event that the init-container script fails to
find the correct clusterrole to use as an ownerref, this commit
means we'll fall back to the previous behaviour of using the
operator deployment.

This is not ideal, because there is a risk that the GC will
delete the webhooks in time, but I think the chance of this
happening is low, and the impact of not having the webhooks is
just that validation will be done in the reconcile loop instead
of immediately (which we already have in the code).

Closes: mhub/qp-planning#5088

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>