feat: Add cluster-wide switch to disable auth (#265)
Adds a switch which determines whether or not Event Streams has
Authentication turned on. Auth is enabled if there is a Kafka
Listeners with authentication enabled.

If the authentication is enabled, then the UI will require you
to sign in. Else, the UI won't require you to sign in. If the
authentication is enabled, then the P2P pods will be secure else
they will have no authentication. Furthermore if authentication is
not enabled, then the automatically configured external endpoints
will have no auth else they will default to the previous auth.
Added a new secure yaml which will allow people to easily test it.

Closes: mhub/qp-planning#5202
Closes: mhub/qp-planning#5203
Closes: mhub/qp-planning#5204
Closes: mhub/qp-planning#5205

Signed-off-by: Julian Goh <julian.goh@ibm.com>