feat: Don't replicate the mm2 object in the EventStreams georep cr (#286)
Initially I thought we'd have a copy of the mm2 spec in the event
streams georeplciator cr.

However, in order for this to be valid, it needs both the connectCluster
filling in and a matching kafkaMirrorMaker2ClusterSpec object in the
list of clusters.

This is tricky to pre-populate in the example cr as it requires the
bootstrap address to be set, and we wouldn't expect this to be known
by the user, nor would we want them filling it in.

The initial plan was to update the ES georep cr from the admin-api
(which we can do) do populate the mm2 spec with all the fields it needs,
but if the user want to set up eg metrics or
additional logging this isn't easily possible until a cluster has
been added using the admin-api.  This is because the metrics
stanza falls under the mm2Spec stanza, and that can't exist until
it has a connectCluster and corresponding connect including the
bootstrap.  Which the user can't be expected to know or configure,
and so is added in by admin-api when adding a pairing.

However this doesn't happen til runtime, and you'd probably be
wanting to set up your metrics at install time...

So, it all got a bit circular. And very data-copying-y.  We are
copying settings back and forth between the esgr cr and the mm2,
which isn't the right thing to do.

Removing mm2 stanza from replicator spec.  If people want to
do custom configuration they should go edit the associated mm2 spec.

Contributes to: mhub/qp-planning#5106

Signed-off-by: Emma Humber <emma.humber@uk.ibm.com>