feat: Endpoint Conditions now return list of conditions (#344)
Endpoint conditions will now return a list of conditions which
will allow us to report all the problems with the current CR
rather than just one. Furthermore, did some refactoring of the
error messages and error reasons so that they are not generic
and they are specific to the error.

Contributes to: mhub/qp-planning#5429

Signed-off-by: Julian Goh <julian.goh@ibm.com>