fix: json bundle (#369)
this commit converts all the yaml files in the olm bundle to json format
before building the bundle. this is done to reduce the size of the
bundle to an acceptable limit. The conversion is temporary only at
build time as it is much clearer to read our csv in yaml format than
one line of json.

Contributes to: mhub/qp-planning#5489

Signed-off-by: Harvey Elsom <harvey.elsom@uk.ibm.com>