fix: Geo-replication messages cleanup (#377)
This commit cleans up various Error messages in the Operator and
the Geo-replicator replicator. Furthermore, the logic of the
georeplicator error messages have been changed to match that of
the Operator's logic. Various files were moved around to improve
the organization of the code so that it is easy to see which
validation belongs to which class.

Closes: mhub/qp-planning#5429

Signed-off-by: Julian Goh <julian.goh@ibm.com>