feat: Add validation of Kafka Connect Spec (#431)
Adds validation of the Kafka Connect spec which allows us to
determine if the customer has passed in the necessary annotations
into the CR which will be used for metering. This validation
checks for missing keys and that the values match
what is expected.

Add KafkaConnectSpecValidation when applying CR but
not to status as ES CR should not have status
about the KafkaConnect components.

Add same validation for annotation check to KafkaConnectS2I
and MirrorMaker2. Also adding new test which checks that
it can pass.

Contributes to: mhub/qp-planning#5527

Signed-off-by: Julian Goh <julian.goh@ibm.com>