fix: Remove custom StatefulSetOperator for schema registry STS (#472)
This commit removes the custom StatefulSetOperator I added
yesterday, replacing it with a simplified implementation based on
subclassing the operator used by Kafka and ZooKeeper.

It's not a perfect fit, as the Kafka and ZK sts operators have
logic built in for custom rolling upgrades that we don't need for
the schema registry. But the amount of useful relevant code we can
inherit outweighs this in my opinion.

I fixed a few values that were missing from the StatefulSetBuilder
I had added. This was resulting in an infinite reconcile loop,
where the operator was trying to remove values from the STS which
were Kubernetes kept restoring.

The StatefulSetDiff method that my new subclassing approach
inherits protects against this as well, as it maintains a list of
values that the operator should ignore and not try to patch.

I also added a check to prevent multiple replicas being deployed
with ephemeral storage, to maintain the behaviour from our Helm
release.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>