fix: Add default examples for all ES Kinds (#485)
The OpenShift UI displays an example from alm-examples in the
editor when someone tries to create a new instance. We don't
have a default example for some of our resource types which
means the editor opens with a broken partial CR.

The new examples exposed a bug in the build_csv.sh script that
prevented it building the correct alm-examples line. I've
corrected the bug and re-run the script to generate the full
correct alm examples.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>