feat: Add ability to create MonitoringDashboard from JSON (#534)
As common services does not have the ability to create a
grafana with persistence enabled, we need to create
this CR to make it work. Hence, added some automation
which should allow us to just specify the json
and the CR would be created.

When the datasource is set to null then the Grafana has no idea
to use prometheus as the default so it will just spit random
data. Hence this commit should fix the automation as it
now sets all datasources to prometheus.

Contributes to: mhub/qp-planning#5761

Signed-off-by: Julian Goh <julian.goh@ibm.com>